Show categories as paths (e.g. cat >> subcat >> subsubcat)
