+++ map-client/src/ListView/ListElement.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 
 import { Poi, PoiWithDistance } from '../types';
 import { Country } from '../state';
-import { presetSingleton } from '../presets';
+import { presetSingleton, getSubcategories } from '../presets';
 import { getAddress, getCuisines, getDistance } from '../info';
 import { OpenStateWrapper } from '../InfoView/OpenState';
 import { PresetIcon } from '../icons';
@@ -26,7 +26,7 @@ const ListElement = (
       </IconDiv>
       <ContentDiv>
         <b>{e.tags['name']}</b><br/>
-        {presetSingleton.getName(e.presetId)}<br/>
+        <Category id={e.presetId} /><br/>
         {getAddress(e, 'long')}<br/>
         Distance: {getDistance(e)}<br/>
         {presetSingleton.hasField(e, 'cuisine') &&
@@ -43,6 +43,30 @@ const ListElement = (
       </ContentDiv>
     </Container>
   );
+};
+
+const Category = ({ id }: { id: string | undefined }) => {
+  if (!id) return null;
+  const subcats = getSubcategories(id);
+  const names = subcats.map(sid => presetSingleton.getName(sid));
+
+  // Format the category like this:
+  //   category >> subcategory >>
+  //   subsubcategory
+  return (
+    <span>
+      <span style={{ color: 'gray' }}>
+        {names.slice(0, -1).join('»')}&raquo;<br />
+      </span>
+      <span>{names[names.length-1]}</span>
+    </span>
+  );
+};
+
+export const getFullLocalName = (id: string, sep: string): string => {
+  const subcats = getSubcategories(id);
+  const names = subcats.map(sid => presetSingleton.getName(sid));
+  return names.join(sep);
 };
 
 const Container = styled.div`
