+++ map-client/src/presets/index.ts
@@ -1,4 +1,5 @@
 export * from './types';
+export * from './utils';
 export * from './presetService';
 export * from './presetNames';
 export * from './presetSearch';
